+++ ordinal-map/src/ordinal/impls/mod.rs
@@ -1,4 +1,5 @@
 mod array;
+mod boxed;
 mod either;
 mod non_zero;
 mod primitive;